requirements.txt
aiogram==3.8.0
httpx==0.27.2
python-dotenv==1.0.1
